Fixes issue with parsing newer rclone output
In an earlier commit, rclone was upgraded from v1.47.0 to v1.64.0.
The newer version produces slightly different file transfer
progress output.

As a result, the motuz UI did not update during a transfer,
but only when the transfer was done.

Our parsing code needed to be modified to
handle the different output.

Closes #440